I'm happy
From one side, I've discovered that sometimes the algorithm can produce infeasible solutions. On the other, not only that's not much a problem since the infeasible solutions are much less than the feasible ones, but also, the C implementation of the code is waaaaaaaaaay faster then the python one (I mean, as you'd expect, but seeing for the first time the difference in performance is quite shocking). This can be a starting point for a heavy testing of the code (right now i just tested it on an instance with 5 boxes, and it went very well