Oh My God
I've been messing with the code all day to understand why the saving of the best local optimum didn't work. Thanks to valgrind i found out that there was an error in the malloc: instead of sizeof(int*) I wrote sizeof(int) for a int** array. I- don't wanna talk about it anymore. I still have to free a lot of things, but at least it now works.