STABLE. Revisit exclude_points
The whole algorithm is finally in a stable state. The only thing to change is the function that cuts out the volume of some extreme points. If those points are internal, I mean, are fully contained by other boxes, their volume can be removed. Otherwise, the function should check in a better way the volume that can be remove (cannot be just all the space on the right, above and in front: only the one we are SURE that cannot be occupied