stable version that can be improved
but for now I'll stop, for real. I have to concentrate on other projects. Still, this version is pretty good, but needs to conclude a TODO, then I can think about improving even further the generation of the unavailable volumes, and finally I'll revisit how to compute the solution for the subset sum problem in its optimization form