My final message
untill the time I'll work on it again, at least. Many ideas, all written in 'cosette', but not enough time to work on them right now. Come on Katema: you will do it, I know it. BELIEVE IN THE ME WHO BELIEVES IN YOU! JUST WHO THE HELL DO YOU THINK WE ARE???